kubectl-gadget: post-process: don't special case first line
The code about filtering out the first line code came from
cmd/kubectl-gadget/bcck8s.go when the bcc tools are directly executed
via the kubectl-exec API on the nodes.

But this package is used for gadgets based on the Trace custom resource
with output in json, so there is no need to edit the header line in that
way.