container-hook: introduce PreCreateContainer events
PreCreateContainer events happen before the container is created and
include the OCI config. The event is included in the "containers"
datasource, so gadgets have an opportunity to inspect the container id
and container config.

The pid, namespaces and other fields are not available since they will
be known only after the container is created.

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>